Usando archivo de utilidades.py para func
