+++ telegramBot/main.py
@@ -2,12 +2,10 @@ import os
 from time import sleep
 import logging
 import pytz
-import requests
 from PIL import Image, ImageDraw, ImageFont
 from prettytable import PrettyTable
 from deta import Deta
 from datetime import datetime
-from bs4 import BeautifulSoup
 from operator import itemgetter
 from io import BytesIO
 from telegram import __version__ as TG_VER
@@ -32,8 +30,6 @@ if __version_info__ < (20, 0, 0, "alpha", 1):
         f"visit https://docs.python-telegram-bot.org/en/v{TG_VER}/examples.html"
     )
 
-# from fastapi import FastAPI, Request
-
 from telegram import (
     ReplyKeyboardMarkup,
     ReplyKeyboardRemove,
@@ -55,7 +51,6 @@ from telegram.ext import (
     ConversationHandler,
     filters,
     CallbackQueryHandler,
-    PicklePersistence,
 )
 
 # Enable logging
@@ -91,232 +86,6 @@ dias_semana = ['lunes', 'martes', 'miercoles', 'jueves', 'viernes', 'sabado', 'd
 meses = ['enero', 'febrero', 'marzo', 'abril', 'mayo', 'junio', 'julio', 'agosto', 'septiembre', 'octubre', 'noviembre','dicembre']
 MARKDOWN_SPECIAL_CHARS = ['_', '*', '[', ']', '(', ')', '~', '`', '>', '#', '+', '-', '=', '|', '{', '}', '.', '!']
 
-# pruebas
-db = deta.Base('simple_db')
-
-def detalle_individual_historico(usuario):
-    mi_historico = dbHistorico.get(usuario)
-    tabla_historico_puntos = PrettyTable()
-    tabla_historico_puntos.title = 'Tabla de puntos obtenidos por carrera'
-    tabla_historico_puntos.field_names = ["Ronda", "Nombre", "Puntos Totales", "Puntos Piloto","Puntos Extras", "Penaizaciones"]
-    tabla_historico_puntos.sortby = "Ronda"
-    puntos_totales = 0
-    for codigo_carrera, puntos in mi_historico['Resultados'].items():
-        carrera = dbCarreras.get(codigo_carrera)
-        puntos_carrera = puntos['normales'] + puntos['extras'] + puntos['penalizaciones']
-        puntos_totales += puntos_carrera
-        tabla_historico_puntos.add_row([ int(carrera['Ronda']), carrera['Nombre'], puntos_carrera, puntos['normales'] , puntos['extras'], puntos['penalizaciones']])
-    texto_abajo = f'Total de puntos: {puntos_totales}'
-    texto_mensaje = f'Este es el detalle de los puntos por carrera que has obtenido hasta el momento'
-    im = Image.new("RGB", (200, 200), "white")
-    dibujo = ImageDraw.Draw(im)
-    letra = ImageFont.truetype("Menlo.ttc", 15)
-    tabladetalletamano = dibujo.multiline_textbbox([0,0],str(tabla_historico_puntos),font=letra)
-    im = im.resize((tabladetalletamano[2] + 20, tabladetalletamano[3] + 40))
-    dibujo = ImageDraw.Draw(im)
-    dibujo.text((10, 10), str(tabla_historico_puntos), font=letra, fill="black")
-    letraabajo = ImageFont.truetype("Menlo.ttc", 10)
-    dibujo.text((20, tabladetalletamano[3] + 20), texto_abajo, font=letra, fill="black")
-    return im, texto_mensaje
-
-def detalle_individual_puntos(usuario):
-    carreras = dbCarreras.fetch([{'Estado':'ARCHIVADA'}, {'Estado':'NO_ENVIADA'}])
-    maximo_horario = datetime.fromisoformat('2023-01-01T00:00:00.000+00:00')
-    ultima_carrera_archivada = {}
-    for carrera in carreras.items:
-        horario_Carrera = datetime.fromisoformat(carrera['Termino'])
-        if(horario_Carrera > maximo_horario):
-            maximo_horario = horario_Carrera
-            ultima_carrera_archivada = carrera
-    tabla_detalle_puntos = PrettyTable()
-    tabla_detalle_puntos.title = ultima_carrera_archivada['Nombre']
-    tabla_detalle_puntos.field_names = ["Pos", "Piloto", "Puntos", "Tus Puntos", "Tus Extras"]
-    tabla_detalle_puntos.sortby = "Pos"
-    resultado_pilotos = dbPuntosPilotos.get(ultima_carrera_archivada['key'])
-    detalles_piloto = dbPilotos.get('2023')['Lista']
-    mi_historico = dbHistorico.get(usuario)
-    mi_lista = mi_historico['Quinielas'][ultima_carrera_archivada['key']]['Lista']
-    mi_quiniela_carrera = mi_historico['Quinielas'][ultima_carrera_archivada['key']]['Carrera']
-    mis_puntos_totales = mi_historico['Resultados'][ultima_carrera_archivada['key']]['extras'] + mi_historico['Resultados'][ultima_carrera_archivada['key']]['normales'] + mi_historico['Resultados'][ultima_carrera_archivada['key']]['penalizaciones']
-    mis_penalizaciones = mi_historico['Resultados'][ultima_carrera_archivada['key']]['penalizaciones']
-    mi_quiniela = mi_historico['Quinielas'][ultima_carrera_archivada['key']]['Lista'].split(',')
-    texto_abajo = f'Tu quiniela: {mi_lista}'
-    texto_mensaje = f'Tus puntos totales fueron: {mis_puntos_totales}, en la imagen puedes ver como los obtuviste. Tambien puedes revisar el documento de las reglas con el comando /ayuda para mas detalles'
-    if mis_penalizaciones < 0:
-        texto_mensaje = f'Tus puntos totales fueron: {mis_puntos_totales}, en la imagen puedes ver como obtuviste los puntos por pilotos y los puntos extras. Estuviste penalizado con {mis_penalizaciones} estos los debes de restar de los puntos de la imagen. Recuerda que puedes revisar las reglas con el comando /ayuda'
-    for numero, resultado in resultado_pilotos['Pilotos'].items():
-        mis_puntos_pilotos = 0
-        mis_puntos_extras = 0
-        if detalles_piloto[numero]['codigo'] in mi_quiniela:
-            mis_puntos_pilotos = resultado['puntos']
-            mi_posicion = mi_quiniela.index(detalles_piloto[numero]['codigo']) + 1
-            if mi_posicion == resultado['posicion']:
-                mis_puntos_extras = 2
-        tabla_detalle_puntos.add_row([resultado['posicion'], detalles_piloto[numero]['codigo'] , resultado['puntos'], mis_puntos_pilotos , mis_puntos_extras])
-    im = Image.new("RGB", (200, 200), "white")
-    dibujo = ImageDraw.Draw(im)
-    letra = ImageFont.truetype("Menlo.ttc", 15)
-    tabladetalletamano = dibujo.multiline_textbbox([0,0],str(tabla_detalle_puntos),font=letra)
-    im = im.resize((tabladetalletamano[2] + 20, tabladetalletamano[3] + 40))
-    dibujo = ImageDraw.Draw(im)
-    dibujo.text((10, 10), str(tabla_detalle_puntos), font=letra, fill="black")
-    letraabajo = ImageFont.truetype("Menlo.ttc", 10)
-    dibujo.text((20, tabladetalletamano[3] + 20), texto_abajo, font=letra, fill="black")    
-    return im, texto_mensaje
-
-def crear_tabla_puntos(obj_carrera):
-    tabla_puntos_piloto = PrettyTable()
-    tabla_puntos_piloto.title = obj_carrera['Nombre']
-    tabla_puntos_piloto.field_names = ["Pos", "Nombre", "Equipo", "Puntos", "Intervalo"]
-    tabla_puntos_piloto.sortby = "Pos"
-    resultado_pilotos = dbPuntosPilotos.get(obj_carrera['key'])
-    detalles_piloto = dbPilotos.get('2023')['Lista']
-    for numero, resultado in resultado_pilotos['Pilotos'].items():
-        tabla_puntos_piloto.add_row([resultado['posicion'], detalles_piloto[numero]['Nombre'] + ' ' + detalles_piloto[numero]['Apellido'], detalles_piloto[numero]['Equipo'], resultado['puntos'], resultado['intervalo']])
-    im = Image.new("RGB", (200, 200), "white")
-    dibujo = ImageDraw.Draw(im)
-    letra = ImageFont.truetype("Menlo.ttc", 15)
-    tablapilotostamano = dibujo.multiline_textbbox([0,0],str(tabla_puntos_piloto),font=letra)
-    im = im.resize((tablapilotostamano[2] + 20, tablapilotostamano[3] + 40))
-    dibujo = ImageDraw.Draw(im)
-    dibujo.text((10, 10), str(tabla_puntos_piloto), font=letra, fill="black")
-    letraabajo = ImageFont.truetype("Menlo.ttc", 10)
-    dibujo.text((20, tablapilotostamano[3] + 20), "Resultados tomados de la pagina oficial de Formula 1", font=letraabajo, fill="black")
-    return im, obj_carrera['Nombre']
-
-def crear_tabla_quinielas(carrera_en_curso, enmascarada=False):
-    """Crear la tabla de las quinielas en una imagen."""
-    carrera_nombre = carrera_en_curso['Nombre']
-    carrera_clave = carrera_en_curso['key']
-    tablaquiniela = PrettyTable()
-    tablaquiniela.title = carrera_nombre
-    tablaquiniela.field_names = ["Fecha/hora", "Nombre", "P1", "P2", "P3", "P4", "P5", "P6", "P7",]
-    tablaquiniela.sortby = "Fecha/hora"
-    datosquiniela = dbQuiniela.fetch({'Carrera':carrera_clave})
-    filas = datosquiniela.items    
-    datos_posiciones = {'P1': [], 'P2': [], 'P3': [], 'P4': [], 'P5': [], 'P6': [], 'P7': []}        
-    for index in range(datosquiniela.count):
-        fila = filas[index]
-        listaquiniela = fila["Lista"].split(",")
-        for pos, piloto in enumerate(listaquiniela):
-            datos_posiciones['P' + str(pos + 1)].append(piloto)
-        if(enmascarada):
-            listaquiniela = ["XXX"] * len(listaquiniela)
-        fechahoraoriginal = datetime.fromisoformat(fila["FechaHora"])
-        fechahoragdl = fechahoraoriginal.astimezone(pytz.timezone('America/Mexico_City'))
-        tablaquiniela.add_row([fechahoragdl.strftime('%Y-%m-%d %H:%M:%S'), fila["Nombre"], listaquiniela[0], listaquiniela[1], listaquiniela[2], listaquiniela[3], listaquiniela[4], listaquiniela[5], listaquiniela[6]])
-    texto_estadisticas = ''
-    if not enmascarada:
-        for pos, datos in datos_posiciones.items():
-            total = len(datos)
-            contador = Counter(datos)
-            texto_estadisticas = texto_estadisticas + pos
-            contador = dict(sorted(contador.items(), key=lambda item: item[1], reverse=True))
-            for piloto, repeticion in contador.items():
-                texto_estadisticas = texto_estadisticas + ' ' + piloto + ' ' + str( round(100 * repeticion/total, 1) ) + '%'
-            texto_estadisticas = texto_estadisticas + '\n'
-    im = Image.new("RGB", (200, 200), "white")
-    dibujo = ImageDraw.Draw(im)
-    letra = ImageFont.truetype("Menlo.ttc", 15)
-    tablaquinielatamano = dibujo.multiline_textbbox([0,0],str(tablaquiniela),font=letra)
-    im = im.resize((tablaquinielatamano[2] + 20, tablaquinielatamano[3] + 40))
-    dibujo = ImageDraw.Draw(im)
-    # dibujo = poner_fondo_gris(dibujo, datosquiniela.count, tablaquinielatamano[2])
-    dibujo.text((10, 10), str(tablaquiniela), font=letra, fill="black")
-    letraabajo = ImageFont.truetype("Menlo.ttc", 10)
-    dibujo.text((20, tablaquinielatamano[3] + 20), "Fecha y hora con el horario de GDL", font=letraabajo, fill="black")
-    return im, carrera_nombre, texto_estadisticas
-
-def crear_tabla_general():
-    tablaresultados = PrettyTable()
-    tablaresultados.title = 'Tabla General Quiniela F1'
-    tablaresultados.field_names = ["Nombre", "Puntos Totales", "Puntos Pilotos", "Puntos Extras", "Penalizaciones"]
-    tablaresultados.sortby = "Puntos Totales"
-    tablaresultados.reversesort = True
-
-    resultados_historicos = dbHistorico.fetch()
-    total_rondas = dbCarreras.fetch([{'Estado':'ARCHIVADA'}, {'Estado':'NO_ENVIADA'}]).count
-    for usuario in resultados_historicos.items:
-        normales = 0
-        extras = 0
-        penalizaciones = 0
-        for carrera in usuario['Resultados']:
-            normales = normales + usuario['Resultados'][carrera]['normales']
-            extras = extras + usuario['Resultados'][carrera]['extras']
-            penalizaciones = penalizaciones + usuario['Resultados'][carrera]['penalizaciones']
-        tablaresultados.add_row([usuario["Nombre"], normales + extras + penalizaciones, normales, extras, penalizaciones])   
-    im = Image.new("RGB", (200, 200), "white")
-    dibujo = ImageDraw.Draw(im)
-    letra = ImageFont.truetype("Menlo.ttc", 15)
-    tablaresultados_tamano = dibujo.multiline_textbbox([0,0],str(tablaresultados),font=letra)
-    im = im.resize((tablaresultados_tamano[2] + 20, tablaresultados_tamano[3] + 40))
-    dibujo = ImageDraw.Draw(im)
-    dibujo = poner_fondo_gris(dibujo, resultados_historicos.count, tablaresultados_tamano[2])
-    dibujo.text((10, 10), str(tablaresultados), font=letra, fill="black")
-    letraabajo = ImageFont.truetype("Menlo.ttc", 10)
-    dibujo.text((20, tablaresultados_tamano[3] + 20), "Total de rondas incluidas: " + str(total_rondas), font=letraabajo, fill="black")
-    return im, total_rondas
-
-def crear_tabla_resultados():
-    carreras = dbCarreras.fetch([{'Estado':'ARCHIVADA'}, {'Estado':'NO_ENVIADA'}])
-    maximo_horario = datetime.fromisoformat('2023-01-01T00:00:00.000+00:00')
-    ultima_carrera_archivada = ''
-    for carrera in carreras.items:
-        horario_Carrera = datetime.fromisoformat(carrera['Termino'])
-        if(horario_Carrera > maximo_horario):
-            maximo_horario = horario_Carrera
-            ultima_carrera_archivada = carrera['key']
-    
-    carrera_codigo = ultima_carrera_archivada
-    carrera_dict = dbCarreras.get(carrera_codigo)
-    carrera_nombre = carrera_dict['Nombre']
-    tablaresultados = PrettyTable()
-    tablaresultados.title = carrera_nombre
-    tablaresultados.field_names = ["Nombre", "Puntos Totales", "Puntos Pilotos", "Puntos Extras", "Penalizaciones"]
-    datosHistoricos = dbHistorico.fetch()
-    usuarios = datosHistoricos.items
-    listaresultados = []
-
-    for index in range(datosHistoricos.count):
-        usuario = usuarios[index]
-        puntos_totales = usuario['Resultados'][carrera_codigo]['normales'] + usuario['Resultados'][carrera_codigo]['extras'] + usuario['Resultados'][carrera_codigo]['penalizaciones']
-        listaresultados.append([usuario["Nombre"], puntos_totales, usuario['Resultados'][carrera_codigo]['normales'], usuario['Resultados'][carrera_codigo]['extras'], usuario['Resultados'][carrera_codigo]['penalizaciones']])
-    listaresultados.sort(key=itemgetter(4,1), reverse=True)
-
-    linea = False
-    ganador = ''
-    puntos_ganador = 0
-    texto_ganador = 'El ganador de la carrera ' + carrera_nombre + ' es: '
-    for index in range(len(listaresultados)):
-        resultado = listaresultados[index]
-        siguiente_resultado = listaresultados[min(index + 1, len(listaresultados) - 1) ]
-        if index == 0:
-            ganador = resultado[0]
-            puntos_ganador = resultado[1]
-        else:
-            if (resultado[1] == puntos_ganador and resultado[4] >= 0):
-                ganador = ganador + ', ' + resultado[0]
-                texto_ganador = 'Los ganadores de la carrera ' + carrera_nombre + ' son: '
-        if(siguiente_resultado[4] < 0 and not linea):
-            tablaresultados.add_row(resultado,  divider=True)
-            linea = True
-        else:
-            tablaresultados.add_row(resultado)
-    if puntos_ganador >= 90:
-        texto_ganador = texto_ganador + ganador + '. Con un total de ' + str(puntos_ganador) + ' puntos.'
-    else:
-        texto_ganador = 'No hubo ganador para la carrera: ' + carrera_nombre + '. Nadie logro hacer 90 o mas puntos. El premio se acumula para la /proxima carrera.'
-    im = Image.new("RGB", (200, 200), "white")
-    dibujo = ImageDraw.Draw(im)
-    letra = ImageFont.truetype("Menlo.ttc", 15)
-    tablaresultados_tamano = dibujo.multiline_textbbox([0,0],str(tablaresultados),font=letra)
-    im = im.resize((tablaresultados_tamano[2] + 20, tablaresultados_tamano[3] + 40))
-    dibujo = ImageDraw.Draw(im)
-    dibujo.text((10, 10), str(tablaresultados), font=letra, fill="black")
-    letraabajo = ImageFont.truetype("Menlo.ttc", 10)
-    dibujo.text((20, tablaresultados_tamano[3] + 20), "Los que tienen penalizaciones no pueden ganar el premio, estan en la segunda seccion de la tabla", font=letraabajo, fill="black")
-    return im, texto_ganador
-
 async def misaldo(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
     usuario = str(update.message.from_user.id)
     pagos_usuario = dbPagos.fetch([{'usuario':usuario, 'estado':'guardado'},{'usuario':usuario, 'estado':'confirmado'}])
@@ -345,7 +114,7 @@ async def start(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
     return ConversationHandler.END
 async def mihistorico(update:Update, context: ContextTypes.DEFAULT_TYPE) -> int:
     """Crear tabla con detalle de puntos por carrera de un participante"""
-    im, mensaje = detalle_individual_historico(str(update.message.from_user.id))
+    im, mensaje = await detalle_individual_historico(str(update.message.from_user.id))
     with BytesIO() as tablaindividualhistoricoimagen:
         im.save(tablaindividualhistoricoimagen, "png")
         tablaindividualhistoricoimagen.seek(0)
@@ -354,7 +123,7 @@ async def mihistorico(update:Update, context: ContextTypes.DEFAULT_TYPE) -> int:
 
 async def mispuntos(update:Update, context: ContextTypes.DEFAULT_TYPE) -> int:
     """Crear tabla con detalle de puntos por participante"""
-    im, mensaje = detalle_individual_puntos(str(update.message.from_user.id))
+    im, mensaje = await detalle_individual_puntos(str(update.message.from_user.id))
     with BytesIO() as tablaindividualpuntosimagen:    
             im.save(tablaindividualpuntosimagen, "png")
             tablaindividualpuntosimagen.seek(0)
@@ -393,6 +162,7 @@ async def pagos(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
     tablapagostamano = dibujo.multiline_textbbox([0,0],str(tablapagos),font=letra)
     im = im.resize((tablapagostamano[2] + 20, tablapagostamano[3] + 40))
     dibujo = ImageDraw.Draw(im)
+    poner_fondo_gris(dibujo=dibujo, total_filas=datosquiniela.count, largo_fila=tablapagostamano[2])
     dibujo.text((10, 10), str(tablapagos), font=letra, fill="black")
     letraabajo = ImageFont.truetype("Menlo.ttc", 10)
     dibujo.text((20, tablapagostamano[3] + 20), "Ronda actual: " + total_carreras, font=letraabajo, fill="black")
@@ -438,9 +208,6 @@ async def proxima(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
 async def cancelar(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
     """Cancels and ends the conversation."""
     user = update.message.from_user
-    # dbprueba = deta.Base('simple_db')
-    # dbprueba.put({'prueba':str(user)})
-    # dbprueba.put({'update':str(update)})
     logger.warning("User %s canceled the conversation.", user.first_name)
     await update.message.reply_text(
         "Comando cancelado, puedes empezar de nuevo con el comando /start", 
@@ -617,7 +384,7 @@ async def quinielas(update:Update, context: ContextTypes.DEFAULT_TYPE) -> int:
         if(ahora > horario_qualy):
             enmascarar = False
             mensaje = "Quinielas para ronda: "
-        im, carrera_nombre, texto_estadisticas = crear_tabla_quinielas(carreras.items[0], enmascarar)
+        im, carrera_nombre, texto_estadisticas = await crear_tabla_quinielas(carreras.items[0], enmascarar)
         with BytesIO() as tablaquinielaimagen:    
             im.save(tablaquinielaimagen, "png")
             tablaquinielaimagen.seek(0)
@@ -629,7 +396,7 @@ async def quinielas(update:Update, context: ContextTypes.DEFAULT_TYPE) -> int:
     return ConversationHandler.END
 
 async def general(update:Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-    im, total_rondas = crear_tabla_general()
+    im, total_rondas = await crear_tabla_general()
     with BytesIO() as tablaresutados_imagen:    
         im.save(tablaresutados_imagen, "png")
         tablaresutados_imagen.seek(0)
@@ -639,7 +406,7 @@ async def general(update:Update, context: ContextTypes.DEFAULT_TYPE) -> int:
 async def resultados(update:Update, context: ContextTypes.DEFAULT_TYPE) -> int:
     """Sends a picture"""
     # aggregar un if, si hay carrera en curso mandar mensaje de espera
-    im, texto = crear_tabla_resultados()
+    im, texto = await crear_tabla_resultados()
     with BytesIO() as tablaresutados_imagen:    
         im.save(tablaresutados_imagen, "png")
         tablaresutados_imagen.seek(0)
@@ -914,78 +681,6 @@ async def p7(update:Update, context:ContextTypes.DEFAULT_TYPE) -> int:
         return P6
     return GUARDAR_PILOTOS
 
-# async def elegir_pilotos(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-#     print("elegir pilotos: ", str(update), str(context.user_data))
-#     query = update.callback_query
-#     await query.answer()
-#     carrera_quiniela = dbCarreras.fetch([{'Estado':'IDLE'}, {'Estado':'EN-CURSO'}])
-#     if query.data == 'NADA':
-#         return ELEGIR_PILOTOS
-#     cuenta_pilotos = len(context.user_data['Lista'])    
-#     if query.data == 'ATRAS':
-#         if cuenta_pilotos > 0:
-#             codigo_eliminado = context.user_data['Lista'].pop()
-#     else:
-#         if cuenta_pilotos < 7:
-#             if not query.data in context.user_data['Lista']:
-#                 context.user_data['Lista'].append(query.data)
-#     pilotoslista = context.user_data['ListaPilotosOrdenada']
-#     COLUMNAS = 5
-#     keyboard = []
-#     cuenta_columnas = COLUMNAS
-#     for pilotonumer in pilotoslista:
-#         pos = ''
-#         if COLUMNAS == cuenta_columnas:
-#             keyboard.append([])
-#             cuenta_columnas = 0
-#         if pilotonumer[1]['codigo'] in context.user_data['Lista']:
-#             p = context.user_data['Lista'].index(pilotonumer[1]['codigo']) + 1
-#             pos = 'P' + str(p) + ' '
-#         keyboard[len(keyboard) - 1].append(InlineKeyboardButton(pos + pilotonumer[1]['codigo'], callback_data=pilotonumer[1]['codigo']))
-#         cuenta_columnas = cuenta_columnas + 1
-#     if len(keyboard[len(keyboard) - 1]) < COLUMNAS:
-#         for i in range(COLUMNAS - len(keyboard[len(keyboard) - 1])):
-#             keyboard[len(keyboard) - 1].append(InlineKeyboardButton(' ', callback_data='NADA'))
-#     cuenta_pilotos = len(context.user_data['Lista'])
-#     if cuenta_pilotos > 0:
-#         keyboard.append([
-#             InlineKeyboardButton("Atras", callback_data='ATRAS'),
-#         ])
-#     if cuenta_pilotos == 7:
-#         keyboard.append([
-#             InlineKeyboardButton("CONFIRMAR", callback_data='CONFIRMAR'),
-#         ])
-#     reply_markup = InlineKeyboardMarkup(keyboard)
-#     texto = 'Asi va tu lista de pilotos (de P1 a P7) para la carrera ' + carrera_quiniela.items[0]['Nombre'] + ':\n' + ",".join(context.user_data['Lista'])
-#     await query.edit_message_text(text=texto, reply_markup=reply_markup)
-#     return ELEGIR_PILOTOS
-
-# async def confirmar_pilotos(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-    # carrera_quiniela = dbCarreras.fetch([{'Estado':'IDLE'}, {'Estado':'EN-CURSO'}])
-    # query = update.callback_query
-    # await query.answer()
-    # if query.data == 'NADA':
-    #     return None
-    # cuenta_pilotos = len(context.user_data['Lista'])    
-    # if query.data == 'ATRAS':
-    #     codigo_eliminado = context.user_data['Lista'].pop()
-    #     return P7
-    # else:
-    #     context.user_data['Lista'].append(query.data)
-    # teclado = []
-    # for fila in update.effective_message.reply_markup.inline_keyboard:
-    #     teclado.append([])
-    #     for boton in fila:
-    #         texto = boton.text
-    #         # if boton.text in context.user_data['Lista']:
-    #         #     texto = 'P7 ' + boton.text
-    #         teclado[len(teclado) - 1].append(InlineKeyboardButton(text= texto, callback_data=boton.callback_data))
-    
-    # reply_markup = InlineKeyboardMarkup(teclado)    
-    # texto = 'Asi va tu lista de pilotos (de P1 a P7) para la carrera ' + carrera_quiniela.items[0]['Nombre'] + ':\n' + ",".join(context.user_data['Lista'])
-    # await query.edit_message_text(text=texto, reply_markup=reply_markup)       
-    # return GUARDAR_PILOTOS
-
 async def guardar_pilotos(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
     query = update.callback_query
     await query.answer()
@@ -1036,9 +731,6 @@ async def main() -> None:
     for pilotonumer in pilotoslista:
         filtropilotos = filtropilotos + '|' + pilotoslista[pilotonumer]['codigo']
     filtropilotos = '^(' + filtropilotos + ')$'
-    # context_types = ContextTypes(context=CustomContext)
-    # Here we set updater to None because we want our custom webhook server to handle the updates
-    # and hence we don't need an Updater instance
     application = (
         Application.builder().token(BOT_TOKEN).updater(None).build()
     )
